Initial commit to link Parent and User
Part of issue: #8 #9

1. We now make use of Django's 'Profile' feature, where we define a
create a Profile model and create a OneToOneField with a User model.
We then create a new ParentProfile every time a new User is created.

The workflow goes like this: Create a User -> (Parent automatically
created) -> Update Parent with other relevant information.

2. All information about the Parent should be exposed in
UserSerializer. It is our entry point.

3. the parent_type field is TEMPORARY. It's just to show how we'll deal
with the serializer. Our ParentProfile will be more complicated once we
link the Child.