Hooked up Child model to Serializers/Views
Part of issue: #11

1. Child has its own view.
2. The User view now has the entire serialized information about the
linked Parent and its children. Though if you want to edit information
about the children, it's better to use the Child view. (It's more REST
that way)
But to GET information, having this entire information serialized in
one place is good, especially since we expect to get all information
of the parent's children if we get the parent.