Framework work on client side. Parent update info.
Part of issue: Not elaborated.

This commit creates a new page for a parent to submit more information
to his/her profile. In addition, a framework is now created on the
client side to handle data.

1. AjaxHelpers does all the Ajax work and will always return a fully
functioning promise.
2. ReactiveStore is not **exactly** a true ReactiveStore. It's actually
just a store for us to store data gotten from the server on the client
side, in memory.
3. In index.jsx, our entire app continually get the parent information
from our ReactiveStore, and update its state if there are new changes.
This is done in 500ms interval for now. We might need to overhaul how
we're doing this in order to increase performance that we may need
later on, but for now, this is fine.

Misc:
1. Implemented a logout functionality.
2. Removed age from our models. It can be calculated from date of birth.
3. Included the new migrations file.